Point int-> double

+++ vezbe/v5-Masinsko ucenje/MasinskoUcenje/MasinskoUcenje/Point.cs
@@ -7,10 +7,10 @@ namespace MasinskoUcenje
 {
     public class Point
     {
-        public int x { get; set; }
-        public int y { get; set; }
+        public double x { get; set; }
+        public double y { get; set; }
 
-        public Point(int x, int y)
+        public Point(double x, double y)
         {
             this.x = x;
             this.y = y;
